Create 3333asn1.cpp
The program takes 2 images, calculates theoptical flow, and uses that to reconstruct a new image. THen finds the mean and standard
deviation of the 2 images.
Make sure openCV is installed in the workspace environment. Compile and run the program with g++.
At the command line, include the path names for the two images to be compared
Example: g++ main.cpp -o run
run 1.png 2.png